+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-# Nightly build 11 (0.2.0) (2020-05-19)
+# Next build 21 (0.2.0) (2020-05-19)
 
 ## Fixes
 
 ### Repo
 - renovate properly opens PRs into nightly branch now ([517b7ac](https://github.com/weeb-cafe/MewChan/commit/517b7ac253927ad6245637b6e52165526907c45b))
+
+### Bot
+- properly re-add mute role when someone rejoins the server ([4ed18d8](https://github.com/weeb-cafe/MewChan/commit/4ed18d85a508f1b3329623c44da1210529fae132))
+- actually exit process if on unhandled rejection if in prod ([b294f14](https://github.com/weeb-cafe/MewChan/commit/b294f14a1ffdef11f763a1680f2d23dfe0b0cb91))
+
+## Chores
+- various major package updates ([6885ba2](https://github.com/weeb-cafe/MewChan/commit/6885ba2cb9855ba685bba5d95ba951e29748f9dc), [31f1916](https://github.com/weeb-cafe/MewChan/commit/31f19165697f6485038d9d8d0ac5fe49d88be2bc), [0615295](https://github.com/weeb-cafe/MewChan/commit/0615295cf5a80973509c4efa28bf278cd1da577c))